+++ src/views/Settings.vue
@@ -44,8 +44,8 @@
                             style="flex-basis: 50%"
                             v-model="darkMode"
                             :label="$t('views.settings.darkModeLabel')"
-                            :messages="$t('views.settings.darkModeMsg')"
                         ></v-switch>
+                        <!-- :messages="$t('views.settings.darkModeMsg')" -->
                         <v-select
                             class="ml-3 mt-3"
                             hide-details
@@ -70,11 +70,12 @@
                             </template>
                         </v-select>
                     </div>
-                    <v-switch
-                        v-model="defaultOpenFavorites"
-                        :label="$t('views.settings.defaultFavorites')"
-                        :messages="$t('views.settings.defaultFavoritesMsg')"
-                    ></v-switch>
+                    <v-select
+                        v-model="defaultOpen"
+                        :items="defaultOpenChoices"
+                        :label="$t('views.settings.defaultPage')"
+                        :messages="$t('views.settings.defaultPageMsg')"
+                    ></v-select>
                     <v-switch
                         v-model="redirectMode"
                         :label="$t('views.settings.redirectModeLabel')"
@@ -156,7 +157,7 @@ export default {
             // "autoplayVideo",
             "scrollMode",
             "hideThumbnail",
-            "defaultOpenFavorites",
+            "defaultOpen",
             "hideCollabStreams",
         ]),
         currentGridSize: {
@@ -224,6 +225,20 @@ export default {
 
             themeId: +localStorage.getItem("theme") || 0,
             themeSet,
+            defaultOpenChoices: Object.freeze([
+                {
+                    text: this.$t("component.mainNav.home"),
+                    value: "home",
+                },
+                {
+                    text: this.$t("component.mainNav.favorites"),
+                    value: "favorites",
+                },
+                {
+                    text: this.$t("component.mainNav.multiview"),
+                    value: "multiview",
+                },
+            ]),
         };
     },
     methods: {